requirements.txt
Django==3.2.4
psycopg2==2.9.1
psycopg2-binary==2.9.1
python3-discogs-client==2.3.12
selenium==3.141.0